+++ bigchaindb/upsert_validator/__init__.py
@@ -1,2 +1,3 @@
 
 from bigchaindb.upsert_validator.validator_election import ValidatorElection # noqa
+from bigchaindb.upsert_validator.validator_election_vote import ValidatorElectionVote # noqa